% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{findInPackages}
\alias{findInPackages}
\title{Find in Packages}
\usage{
findInPackages(needle, ignore.case = TRUE, ...,
  lookInPackagename = TRUE, lookInDescription = TRUE,
  lookInHelp = TRUE, lookInData = TRUE, showOnlyContaining = TRUE)
}
\arguments{
\item{needle}{character string containing a string or regular expression specifying want you to search for}

\item{ignore.case}{logical. if \code{FALSE} the pattern matching is \emph{case sensitive}}

\item{...}{additional parameters for \code{\link[base:grep]{grep}}}

\item{lookInPackagename}{logical. if \code{FALSE} the packagenames won't be checked}

\item{lookInDescription}{logical. if \code{FALSE} the package descriptions won't be checked}

\item{lookInHelp}{logical. if \code{FALSE} the package help won't be checked}

\item{lookInData}{logical. if \code{FALSE} the data in the packages won't be checked}

\item{showOnlyContaining}{logical. if \code{TRUE} only packages that have one match will be displayed}
}
\value{
\code{findInPackages("mtcars")} returns a table with all Packages that match for the word \emph{mtcars} in either the packagename, description, help or data
}
\description{
Function that allows you to search in all installed packages for a specific string
}
\examples{
findInPackages("mtcars")
  Package  Packagename Description Help  Data
1 datasets FALSE       FALSE       TRUE  TRUE

}
